requirements.txt
Flask==3.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
openai==1.3.8
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0